+++ src/cells/_Cell.ts
@@ -6,15 +6,13 @@ import { prefix } from "../enums";
 import { profile } from "../profiler/decorator";
 import type { HiveCells } from "../Hive";
 
-
 @profile
 export abstract class Cell {
+  private readonly hive: Hive;
+  private readonly ref: string;
+  private master: Master | undefined;
 
-  readonly hive: Hive;
-  readonly ref: string;
-  master: Master | undefined;
-
-  constructor(hive: Hive, cellName: string) {
+  private constructor(hive: Hive, cellName: string) {
     this.hive = hive;
     this.ref = cellName;
 
@@ -25,64 +23,81 @@ export abstract class Cell {
       this.master = Apiary.masters[prefix.master + this.ref];
   }
 
-  get refCache() {
+  private get refCache() {
     return this.ref.split("_")[0];
   }
 
-  get pos(): RoomPosition {
+  private get pos(): RoomPosition {
     return this.hive.pos;
   }
 
-  delete() {
-    for (let cellType in this.hive.cells)
-      if (this.hive.cells[<keyof HiveCells>cellType]!.ref === this.ref)
-        delete this.hive.cells[<keyof HiveCells>cellType];
-    if (this.master)
-      this.master.delete();
+  private delete() {
+    for (const cellType in this.hive.cells)
+      if (this.hive.cells[cellType as keyof HiveCells]!.ref === this.ref)
+        delete this.hive.cells[cellType as keyof HiveCells];
+    if (this.master) this.master.delete();
   }
 
   // first stage of decision making like do i a logistic transfer do i need more masters
-  update<K extends keyof this>(updateMapKey: K[] = [], nonforceMapKey: K[] = []): void {
+  private update<K extends keyof this>(
+    updateMapKey: K[] = [],
+    nonforceMapKey: K[] = []
+  ): void {
     // updating structure object to actual data
 
     _.forEach(Object.keys(this), (key: K) => {
-      let data = this[key];
-      if (data instanceof Structure || data instanceof Source || data instanceof Mineral) {
-        let gameObject = Game.getObjectById(data.id);
+      const data = this[key];
+      if (
+        data instanceof Structure ||
+        data instanceof Source ||
+        data instanceof Mineral
+      ) {
+        const gameObject = Game.getObjectById(data.id);
         if (gameObject || !nonforceMapKey.includes(key))
-          this[key] = <typeof data>gameObject;
+          this[key] = gameObject as typeof data;
       }
     });
 
     _.forEach(updateMapKey, (key: K) => {
       for (const inMap in this[key]) {
-        let data = this[key][inMap];
-        let gameObject = Game.getObjectById(inMap);
-        if (gameObject)
-          this[key][inMap] = <typeof data>gameObject;
-        else
-          delete this[key][inMap];
+        const data = this[key][inMap];
+        const gameObject = Game.getObjectById(inMap);
+        if (gameObject) this[key][inMap] = gameObject as typeof data;
+        else delete this[key][inMap];
       }
     });
   }
 
   // second stage of decision making like where do i need to spawn creeps or do i need
-  abstract run(): void;
+  protected abstract run(): void;
 
-  setCahe<K extends keyof this, T extends this[K]>(key: K, baseValue: T) {
-    if (!(<string>key in Memory.cache.hives[this.hive.roomName].cells[this.refCache]))
-      Memory.cache.hives[this.hive.roomName].cells[this.refCache][<string>key] = baseValue;
+  private setCahe<K extends keyof this, T extends this[K]>(
+    key: K,
+    baseValue: T
+  ) {
+    if (
+      !(
+        (key as string) in
+        Memory.cache.hives[this.hive.roomName].cells[this.refCache]
+      )
+    )
+      Memory.cache.hives[this.hive.roomName].cells[this.refCache][
+        key as string
+      ] = baseValue;
   }
 
-  toCache<K extends keyof this, T extends this[K]>(key: K, value: T) {
-    Memory.cache.hives[this.hive.roomName].cells[this.refCache][<string>key] = value;
+  private toCache<K extends keyof this, T extends this[K]>(key: K, value: T) {
+    Memory.cache.hives[this.hive.roomName].cells[this.refCache][key as string] =
+      value;
   }
 
-  fromCache<K extends keyof this, T extends this[K]>(key: K) {
-    return <T>Memory.cache.hives[this.hive.roomName].cells[this.refCache][<string>key];
+  private fromCache<K extends keyof this, T extends this[K]>(key: K) {
+    return Memory.cache.hives[this.hive.roomName].cells[this.refCache][
+      key as string
+    ] as T;
   }
 
-  get print(): string {
+  private get print(): string {
     return `<a href=#!/room/${Game.shard.name}/${this.pos.roomName}>["${this.ref}"]</a>`;
   }
 }
